risc0_core
Module field

Finite field types and operations

Defines base fields and extension fields used for finite field-based operations across the RISC Zero zkVM architecture

Modules

baby_bear

Baby bear field.

goldilocks

Goldilocks field.

Traits

Elem

Subfield elements that can be compared, copied, and operated on via multiplication, addition, and subtraction

ExtElem

A field extension which can be constructed from a subfield element Elem

Field

A pair of fields, one of which is an extension field of the other.

RootsOfUnity

Roots of unity for the field whose elements are represented by ExtElem and whose subfield elements are represented by Elem

Functions

map_pow

Equivalent to exponents.map(|exponent| base.pow(exponent)).collect(), but optimized to execute fewer multiplies. Exponents must be sorted and strictly increasing.