risc0_core::field
Trait ExtElem

pub trait ExtElem:
    Elem
    + From<Self::SubElem>
    + Neg<Output = Self>
    + PartialEq
    + Eq
    + Add<Self::SubElem, Output = Self>
    + Sub<Self::SubElem, Output = Self>
    + Mul<Self::SubElem, Output = Self>
    + AddAssign<Self::SubElem>
    + SubAssign<Self::SubElem>
    + MulAssign<Self::SubElem> {
    type SubElem: Elem + Add<Self, Output = Self> + Sub<Self, Output = Self> + Mul<Self, Output = Self>;

    const EXT_SIZE: usize;

    // Required methods
    fn from_subfield(elem: &Self::SubElem) -> Self;
    fn from_subelems(elems: impl IntoIterator<Item = Self::SubElem>) -> Self;
    fn subelems(&self) -> &[Self::SubElem];
}

A field extension which can be constructed from a subfield element Elem

Represents an element of an extension field. This extension field is associated with a base field (sometimes called “subfield”) whose element type is given by the generic type parameter.

Required Associated Constants

const EXT_SIZE: usize

The degree of the field extension

This the degree of the extension field when interpreted as a vector space over the base field. Thus, an ExtElem can be represented as EXT_SIZE SubElems.

Required Associated Types

type SubElem: Elem + Add<Self, Output = Self> + Sub<Self, Output = Self> + Mul<Self, Output = Self>

An element of the base field

This type represents an element of the base field (sometimes called “subfield”) of this extension field.

Required Methods

fn from_subfield(elem: &Self::SubElem) -> Self

Interpret a base field element as an extension field element

Every SubElem is (mathematically) an ExtElem. This constructs the ExtElem equal to the given SubElem.

fn from_subelems(elems: impl IntoIterator<Item = Self::SubElem>) -> Self

Construct an extension field element

Construct an extension field element from a (mathematical) vector of SubElems. This vector is length EXT_SIZE.

fn subelems(&self) -> &[Self::SubElem]

Express an extension field element in terms of base field elements

Returns the (mathematical) vector of SubElems equal to the ExtElem. This vector is length EXT_SIZE.

Object Safety

This trait is not object safe.